requirements.txt
streamlit
yfinance
newsapi-python
nltk
tensorflow
pandas
numpy
scikit-learn
